Update unit tests for coverage and resolve any failures
