+++ tests/images/airflow/2.11.0/python/mwaa/logging/test_cloudwatch_handler_2_11_0.py
@@ -38,7 +38,7 @@ def mock_watchtower():
 
 @pytest.fixture
 def mock_fluent():
-    with patch('mwaa.logging.cloudwatch_handlers.fluent_handler.FluentHandler') as mock:
+    with patch('mwaa.logging.fork_safe_handler.ForkSafeFluentHandler') as mock:
         yield mock
 
 @pytest.fixture(autouse=True)
